begin adding custom enemy
This is the beginning of adding the flying rat as a custom enemy; I believe I understand how its actually being used at this point and I managed to get the model to load. I am going to remove Algernon's special NPC here to leave it as just a flying rat enemy for now with NO miniboss tied to it; that way I can keep it as simple as possible with nothing conflicting.

monster_rat.cpp will also be edited to match its original state in the next commit.

added an important note section to the TODO.md for certain things, like how the MODELS NEED TO BE IN SLAB format to be exported from Magicalvox or w/e its called.

Monster.hpp has been edited to allow the creation of monster_flyingrat.cpp