SpellTesting - Managed to make a spell
It works! A lot of stuff needed to be changed for it to work; But I have a much greater understanding for spells as a whole! Now is a matter of making a new, "PROPER" spell for the Cleric... Eesh... That'll be a new spell called "Poison Flame", and it'd essentially do exactly what it says. Apply Poison, and Flame, at the same time. I want to try and see if it's possible to invert the "healing range" to be used in place so it can apply affects to an AOE of multiple enemies at once.